requirements.txt
requests>=2.31.0
Pillow>=10.0.0
huggingface-hub>=0.20.0
